Tests/LibWeb/Text/expected/css/custom-property-registration-invalidation.txt

before @property color: rgb(255, 0, 0)
after @property color: rgb(0, 128, 0)
after removing @property color: rgb(255, 0, 0)
before registerProperty color: rgb(255, 0, 0)
after registerProperty color: rgb(0, 0, 255)
